backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cnocr==2.3.2.1
cnstd==1.2.6.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dashscope==1.24.3
datasets==4.0.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
doclayout_yolo==0.0.4
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.9
kombu==5.5.4
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
openai==1.106.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
optimum==1.27.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pix2text==1.1.4
platformdirs==4.4.0
polars==1.32.3
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
PyMuPDF==1.26.4
pyparsing==3.2.3
pyreadline3==3.5.4
pyspellchecker==0.8.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.4
pytz==2025.2
PyYAML==6.0.2
rapidocr==3.3.1
redis==6.4.0
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.1
seaborn==0.13.2
sentry-sdk==2.35.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.3
stringzilla==3.12.6
sympy==1.14.0
thop==0.1.1.post2209072238
tokenizers==0.21.4
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
transformers==4.53.3
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.189
ultralytics-thop==2.0.16
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wandb==0.21.3
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1

# ==============================================================================
# Feature 1-4: 个性化学习系统依赖 (V23.0)
# ==============================================================================
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.0.7  # Feature 4: PDF生成
python-docx==1.1.0  # Feature 4: Word文档生成（备选）